Redesign text layout shaping. (#2)
This PR redesigns how text layouts are shaped. It primarily changes how the
vertical and horizontal bases of boxes are calculated. This allows for more
options to choose from where defining where the base should be and adds
explicit anchor items that can be referred to when defining the base.